@@ -26,6 +26,7 @@
     "mongoose": "^5.11.18",
     "multer": "^1.4.2",
     "multer-storage-cloudinary": "^4.0.0",
+    "node-macaddress": "^0.2.4",
     "nodemailer": "^6.6.3",
     "public-ip": "^4.0.4",
     "redis": "^3.1.2",
@@ -58,6 +59,6 @@
     "@types/redis": "^2.8.28",
     "@types/uuid": "^8.3.0",
     "nodemon": "^2.0.7",
-    "typescript": "^4.3.4"
+    "typescript": "^4.4.2"
   }
 }
