@@ -87,9 +87,7 @@ class CartEntity extends BaseEntity {
             uuid: this.uuid,
             quantity: this.quantity,
             product: this.product,
-            created_at: this.created_at,
-            updated_at: this.updated_at,
-            deleted_at: this.deleted_at,
+
         };
     }
 
