@@ -1,6 +1,7 @@
-import { PaymentMethod, StatusPayment } from "../enums/enum";
+import { PaymentMethod, StatusOrder, StatusPayment } from "../enums/enum";
 import { IProduct } from "./cart";
 import { IEmbed } from "./product";
+
 
 export interface IPaymentEntity {
     uuid: string
@@ -15,11 +16,13 @@ export interface IPaymentEntity {
     idempotency: string
     created_by: IEmbed
     seller_by: IEmbed
-    status: StatusPayment
+    status_payment: StatusPayment
+    status: StatusOrder
     card_name: string
     quantity: number
     product: IProduct
     total_price: number
+    macaddress: string | null
     image: string
     cloudinary_id: string
     created_at: Date | null
