@@ -6,6 +6,6 @@ export interface ICartRepository {
     update(data: CartEntity): Promise<{ success: true }>
     findOne(uuid: string, user_uuid: string): Promise<CartEntity | null>
     findOneMyCart(user_uuid: string): Promise<CartEntity | null>
-    findAll(user_uuid: string): Promise<CartEntity[]>
+    findAll(user_uuid: string): Promise<CartEntity | null>
     delete(uuid: string, user_uuid: string): Promise<{ success: true }>
 }
