@@ -4,7 +4,7 @@ import { IUser } from "../models/interfaces/user";
 import { TYPES } from "../types";
 import { EventDispatcher } from "event-dispatch";
 import { v4 as uuid } from 'uuid'
-import { ErrorNotFound } from "../helpers/errors";
+import { ErrorBadRequest, ErrorNotFound } from "../helpers/errors";
 import { ISellerRequestService } from "./interfaces/sellerRequest";
 import CreateSellerRequest from "../request/sellerRequest/createSellerRequest";
 import SellerRequestEntity from "../entities/sellerRequest";
@@ -17,6 +17,10 @@ import ProfileEntity from "../entities/profile";
 import { UserRole } from "../entities/enums/enum";
 import UserEntity from "../entities/user";
 import { IBankService } from "./interfaces/bank";
+import { IProvinceService } from "./interfaces/province";
+import { ICityService } from "./interfaces/city";
+import { IDistrictService } from "./interfaces/district";
+import { IVillageSerivce } from "./interfaces/village";
 
 @injectable()
 class SellerRequestService implements ISellerRequestService {
@@ -25,11 +29,24 @@ class SellerRequestService implements ISellerRequestService {
         @inject(TYPES.ProfileService) private profileService: IProfileService,
         @inject(TYPES.UserService) private userService: IUserService,
         @inject(TYPES.BankService) private bankService: IBankService,
-        @inject(TYPES.ProducerDispatcher) private dispatcher: EventDispatcher,
+        @inject(TYPES.ProvinceService) private provinceService: IProvinceService,
+        @inject(TYPES.CityService) private cityService: ICityService,
+        @inject(TYPES.DistrictService) private districtService: IDistrictService,
+        @inject(TYPES.VillageService) private villageService: IVillageSerivce,
     ) { }
     async create(data: CreateSellerRequest, user: IUser): Promise<{ success: true }> {
         const searchUser = await this.profileService.findOne(user.uuid)
-        if (!searchUser) throw new ErrorNotFound('User not found', '@Service Create Seller')
+        if (!searchUser) throw new ErrorNotFound('User tidak ditemukan', '@Service Create Seller')
+        const sellerRequest = await this.sellerRequestRepository.findOneByUserUuid(searchUser.created_by.uuid ?? '')
+        if (sellerRequest) throw new ErrorBadRequest('Kamu sudah pernah request sebagai seller', '@Service Create Request Seller')
+        const province = await this.provinceService.findOne(data.province_uuid)
+        if (!province) throw new ErrorNotFound('Province tidak ditemukan', '@Service Create Request Seller')
+        const city = await this.cityService.findOne(data.city_uuid)
+        if (!city) throw new ErrorNotFound('City tidak ditemukan', '@Service Create Request Seller')
+        const district = await this.districtService.findOne(data.district_uuid)
+        if (!district) throw new ErrorNotFound('District tidak ditemukan', '@Service Create Request Seller')
+        const village = await this.villageService.findOne(data.village_uuid)
+        if (!village) throw new ErrorNotFound('Village tidak ditemukan', '@Service Create Request Seller')
         const bank = await this.bankService.findOne(data.bank_uuid)
         const entity = new SellerRequestEntity({
             uuid: uuid(),
@@ -46,7 +63,23 @@ class SellerRequestService implements ISellerRequestService {
             ktp_image: data.ktp_image,
             name: data.name,
             status: 'Pending',
-            bank: bank.toJson(),
+            bank: bank,
+            province: {
+                uuid: province.uuid,
+                name: province.name,
+            },
+            city: {
+                uuid: city.uuid,
+                name: city.name,
+            },
+            district: {
+                uuid: district.uuid,
+                name: district.name
+            },
+            village: {
+                uuid: village.uuid,
+                name: village.name,
+            },
             deleted_at: null,
             created_at: new Date,
             updated_at: new Date,
@@ -70,6 +103,10 @@ class SellerRequestService implements ISellerRequestService {
         if (!searchUser) throw new ErrorNotFound('User not found', '@service update to seller')
         const bank = await this.bankService.findOne(data.bank_uuid)
         if (!bank) throw new ErrorNotFound('Bank not found', '@Service update to seller')
+        const province = await this.provinceService.findOne(data.province_uuid)
+        const city = await this.cityService.findOne(data.city_uuid)
+        const district = await this.districtService.findOne(data.district_uuid)
+        const village = await this.villageService.findOne(data.village_uuid)
         const entity = new SellerRequestEntity({
             uuid: searchSellerRequest?.uuid ?? '',
             card_holder_name: data.card_holder_name,
@@ -87,6 +124,10 @@ class SellerRequestService implements ISellerRequestService {
             status: 'Pending',
             bank: bank,
             deleted_at: null,
+            province: province,
+            city: city,
+            district: district,
+            village: village,
             created_at: searchSellerRequest?.created_at ?? new Date,
             updated_at: searchSellerRequest?.updated_at ?? new Date,
         })
@@ -96,7 +137,7 @@ class SellerRequestService implements ISellerRequestService {
 
     async updateToSeller(uuid: string): Promise<{ success: true }> {
         const searchRequest = await this.sellerRequestRepository.findOne(uuid)
-        // if (searchRequest?.status === 'Completed') throw new ErrorNotFound('Expired Data', '@Servoce change to seller ')
+        if (searchRequest?.status === 'Completed') throw new ErrorNotFound('Expired Data', '@Servoce change to seller ')
         if (!searchRequest) throw new ErrorNotFound('Data not found', '@Service Change member to seller')
         const searchProfile = await this.profileService.findOne(searchRequest.created_by.uuid ?? '')
         if (!searchProfile) throw new ErrorNotFound('Data not found', '@Service Change user to seller in Profile')
@@ -106,10 +147,10 @@ class SellerRequestService implements ISellerRequestService {
         const profileEntity = new ProfileEntity({
             uuid: searchProfile.uuid,
             card_number: searchRequest.card_number,
-            province: searchProfile.province,
-            city: searchProfile.city,
-            district: searchProfile.district,
-            village: searchProfile.village,
+            province: searchRequest.province,
+            city: searchRequest.city,
+            district: searchRequest.district,
+            village: searchRequest.village,
             cloudinary_id: searchProfile.cloudinary_id,
             email: searchProfile.email,
             created_by: searchProfile.created_by,
@@ -119,6 +160,7 @@ class SellerRequestService implements ISellerRequestService {
             updated_at: searchProfile.updated_at,
             deleted_at: searchProfile.deleted_at,
             roles: [UserRole.MEMBER, UserRole.Seller],
+            is_active: searchProfile.is_active,
             slug: searchProfile.slug,
             address: searchProfile.address,
             bank: searchRequest.bank
@@ -139,6 +181,7 @@ class SellerRequestService implements ISellerRequestService {
         await this.userService.updateForSeller(userEntity)
         await this.profileService.updateForSeller(profileEntity)
         searchRequest.status = 'Completed'
+        searchRequest.deleted_at = new Date()
         await this.sellerRequestRepository.update(searchRequest)
 
         return { success: true }
