@@ -9,8 +9,8 @@ const TYPES = {
     UserController: Symbol.for("UserController"),
     UserService: Symbol.for("UserService"),
     UserRepository: Symbol.for("UserRepository"),
-    // ProfileRouter: Symbol.for("ProfileRouter"),
-    // ProfileController: Symbol.for("ProfileController"),
+    ProfileRouter: Symbol.for("ProfileRouter"),
+    ProfileController: Symbol.for("ProfileController"),
     ProfileService: Symbol.for("ProfileService"),
     ProfileRepository: Symbol.for("ProfileRepository"),
     CommentRouter: Symbol.for("CommentRouter"),
